pyproject.toml
[tool.poetry]
name = "manong-style"
version = "0.1.0"
description = ""
authors = ["Guan Hao <raptium@gmail.com>"]
packages = [
    { include = "manong" },
]

[tool.poetry.scripts]
mns = 'manong.cli:entry'

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^6.2.1"
numpy = "^1.18.0"
click = "^7.0"
PyYAML = "^5.2"

[tool.poetry.dev-dependencies]
yapf = "^0.29.0"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
